Break blended fonts out into subtabs by script
